Fire alarms should reliably open doors when reset
So I think I figured it out: There's spam protection on the alarm side, but it is only set and checked by the alarm. Now it is set for the alarm by the reset, but the reset does not check for it.

This may have unforeseen gameplay implications! They are almost certainly not worse than trying to open a fukken fire door while you're burning alive and having it do the two-tick two-step on you, opening and closing without rhyme or reason.

Enjoy your death trap, ladies!